requirements.txt
flask
flask-cors
pandas
seaborn
matplotlib
azure-storage-blob
gunicorn
auth0-server-python>=1.0.0b7
flask[async]>=2.0.0
python-dotenv>=1.0.0